Separate items page from its popup. Encapsulate SelectedCount, because it is a view component
