+++ Wasted/Pages/Deals/OrderPopup.xaml
@@ -1,0 +1,36 @@
+﻿<?xml version="1.0" encoding="UTF-8" ?>
+<pages:PopupPage
+	xmlns:pages="clr-namespace:Rg.Plugins.Popup.Pages;assembly=Rg.Plugins.Popup"
+	xmlns="http://xamarin.com/schemas/2014/forms"
+	xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
+    x:Class="Wasted.Pages.Deals.OrderPopup">
+
+    <!--#region Order view-->
+    <StackLayout x:Name="orderView" Style="{StaticResource DealCardViewLayout}"
+                 VerticalOptions="Center">
+        <StackLayout BackgroundColor="White">
+
+            <StackLayout Orientation="Horizontal"
+                    HorizontalOptions="Center"
+                    Spacing="20">
+                <Label x:Name="countLabel" Style="{StaticResource RatingLabel}"
+                        Text="{Binding SelectedCount, StringFormat='Count: {0}'}" />
+
+
+                <Stepper x:Name="stepper"
+                        Maximum="{Binding Quantity}"
+                        Increment="1"
+                        ValueChanged="StepperValueChanged"/>
+
+            </StackLayout>
+
+            <StackLayout Style="{StaticResource ConfirmCancelLayout}">
+                <Button Style="{StaticResource CancelButton}" Clicked="OnCancelClicked"  WidthRequest="75"/>
+                <Button Style="{StaticResource ConfirmButton}" Clicked="OnConfirmClicked" WidthRequest="220"/>
+            </StackLayout>
+
+        </StackLayout>
+    </StackLayout>
+    <!--#endregion-->
+    
+</pages:PopupPage>
